Management of errors & GUI
Before this version, when an error was encountered while opening a Word document, the script crashed. Now it just ignores this erronous file and continue. The erronous files are listed at the end of the program.

New graphic interface more clear for the user thanks to the usage of main title